Update styles.css
Header font size changed from 2em to 1em.
In CSS (Cascading Style Sheets), em is a unit of measurement which indicates size relative to the size of the font. 1em means "equal to the current font size", and 2em means 2 times the current font size. (For example in a 2em unit of measurement, if your font is 11px (11 pixels), then 2em equals 22 pixels.)